Initial plumbing for dynamic cmdr cmd loading
